PageWithTable: fix error on drill-down to JS page with search required
If a JS table page is marked with searchRequired=true and the user
activates it with a double click on the corresponding row in the parent
node's detail table, an error occurred in _resetTableData(), because
the detail table was not yet created. This has been fixed by calling
ensureDetailTable() before calling _resetTableData().

457043

+++ eclipse-scout-core/src/desktop/outline/pages/PageWithTable.ts
@@ -223,7 +223,18 @@ export class PageWithTable extends Page implements PageWithTableModel {
     this._resetTableData();
   }
 
+  /**
+   * Reverts the data in the {@link detailTable} to the original state by reloading it. For tables with
+   * {@link searchRequired} = true, the data is not reloaded. Instead, the old rows are deleted, a table
+   * status is displayed and {@link searchFilterCompleted} is set to false.
+   *
+   * This method must only be called after the detail table has been created ({@link ensureDetailTable}).
+   */
   protected _resetTableData() {
+    // It's allowed to have no table - but we don't have to load data in that case
+    if (!this.detailTable) {
+      return;
+    }
     if (this.searchRequired) {
       this.detailTable.deleteAllRows();
       this.setSearchFilterCompleted(false);
@@ -265,6 +276,8 @@ export class PageWithTable extends Page implements PageWithTableModel {
 
   override ensureLoadChildren(): JQuery.Promise<any> {
     if (this.searchRequired && !this.searchFilterCompleted) {
+      // Show table status
+      this.ensureDetailTable();
       this._resetTableData();
       return $.resolvedPromise();
     }
